✅  split id and uuid tests
